docker_compose/movies_admin/requirements/requirements-dev.txt
-r requirements.txt
flake8==6.0.0
isort==5.12.0
django-cors-headers==3.14.0
django_debug_toolbar==3.8.1
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1